Update AbstractApi.php
I had a problem with caching none 200 code responses, mostly with the "currentGame" request. Every Response was cached, even the 503 oder 404 ones, with the amount of seconds i specified.

The code above checks if the returned content is a Response object. If it is it will check the setted code for a 200 response and only this ones will be cached.
If the content is not a Resonse object it will be cached without restriction